requirements.txt
pandas>=2.0.0
numpy>=1.24.0
xgboost>=2.0.0
ccxt>=4.0.0
pyyaml>=6.0
scikit-learn>=1.3.0
